+++ src/test/resources/extend.entity.ts
@@ -1,20 +1,23 @@
-import { ApiModelProperty, ApiModel } from "../../lib/swagger-express-ts";
-import { MainEntity } from "./main.entity";
+import {
+    ApiModelProperty,
+    ApiModel,
+} from '../../../node_modules/swagger-express-ts';
+import { MainEntity } from './main.entity';
 
 @ApiModel({
-  description: "Extend description",
-  name: "Extend"
+    description: 'Extend description',
+    name: 'Extend',
 })
 export class ExtendingEntity extends MainEntity {
-  @ApiModelProperty({
-    description: "",
-    required: true
-  })
-  extendNumber: string;
+    @ApiModelProperty({
+        description: '',
+        required: true,
+    })
+    public extendNumber: string;
 
-  @ApiModelProperty({
-    description: "extend main name",
-    required: true
-  })
-  mainNumber: string;
+    @ApiModelProperty({
+        description: 'extend main name',
+        required: true,
+    })
+    public mainNumber: string;
 }